# golang-gin/api/reservation/swagger.yaml
definitions:
  controller.airportDto:
    properties:
      code:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  controller.ticketDto:
    properties:
      flight_id:
        type: string
      id:
        type: string
    type: object
info:
  contact: {}
paths:
  /health/check:
    get:
      consumes:
      - application/json
      description: do health check
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              code:
                type: string
            type: object
      summary: health check
      tags:
      - health
  /health/ready:
    get:
      consumes:
      - application/json
      description: do health ready
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              code:
                type: string
            type: object
      summary: health ready
      tags:
      - health
  /v1/airports:
    get:
      description: list airports desc
      parameters:
      - in: query
        name: cursor_next
        type: integer
      - in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              code:
                type: string
              items:
                $ref: '#/definitions/controller.airportDto'
            type: object
      summary: list airports
      tags:
      - airport
  /v1/flights/{flight_id}/tickets/optimistic:
    post:
      description: create a ticket using optimistic lock desc
      parameters:
      - description: Flight ID
        in: path
        name: flight_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              code:
                type: string
              item:
                $ref: '#/definitions/controller.ticketDto'
            type: object
      summary: create a ticket
      tags:
      - ticket
  /v1/flights/{flight_id}/tickets/pessimistic:
    post:
      description: create a ticket using pessimistic lock desc
      parameters:
      - description: Flight ID
        in: path
        name: flight_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              code:
                type: string
              item:
                $ref: '#/definitions/controller.ticketDto'
            type: object
      summary: create a ticket
      tags:
      - ticket
swagger: "2.0"